% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_quant.R
\name{cut_quant}
\alias{cut_quant}
\title{Create strata based on quantiles of a score}
\usage{
cut_quant(v, q, int = TRUE)
}
\arguments{
\item{v}{Vector of scores (typically propensity scores)}

\item{q}{Vector of desired quantiles (between 0 and 1) at which to cut the strata}

\item{int}{Boolean whether to return strata as integers. Default is `TRUE`}
}
\value{
Vector of strata
}
\description{
Create strata based on quantiles of a score
}
\examples{
cut_quant(1:9, c(1/3, 2/3), int = TRUE)
}
